+++ lib/active_interaction/errors.rb
@@ -1,6 +1,7 @@
 module ActiveInteraction
   class InteractionInvalid < ::StandardError; end
   class InvalidDefaultValue < ::StandardError; end
+  class InvalidNestedValue < ::StandardError; end
   class InvalidValue < ::StandardError; end
   class MissingValue < ::StandardError; end
 end