+++ lib/active_interaction/filters/hash_filter.rb
@@ -25,7 +25,7 @@ module ActiveInteraction
     def self.prepare(key, value, options = {}, &block)
       case value
         when Hash
-          convert_values(value.dup, &block)
+          convert_values(value.merge(options[:default] || {}), &block)
         else
           super
       end
@@ -41,6 +41,8 @@ module ActiveInteraction
       end
 
       hash
+    rescue InvalidValue, MissingValue
+      raise InvalidNestedValue
     end
     private_class_method :convert_values
   end